Add files via upload
The file shipsites.ssc should be placed in your tworings folder.  When you are in Celestia and the focus is on Earth, then you can go to any of these positions, e.g., <enter> JonesObs1 <enter>  to go to the first observation as they are numbered on Jones's list.  These are only positions, not times.  Note also once you have "arrived" you will find yourself anomalously oriented even if correctly positioned, and you may have to ascend (1) or cartwheel or somersault (cursor arrows) to orient yourself as an observer at the surface of the Earth facing East or West.  Further you will have to set the time correctly.