Add files via upload
This version of the Extra for two rings includes Jones's ship sites, Cassini's observations, and the needed jpl ephemerides to look at Cassini's dates.  There is also a spare version of the ephemerides for 1800 to 2100 in case the user experiments with Cassini's dates and loses the 1800 to 2100 ephemerides in doing that.